power: pmic: sunxi: add AXP717 SPL driver
On boards using the AXP717 PMIC, the DRAM rail is often not setup
correctly at reset time, so we have to program the PMIC very early in
the SPL, before running the DRAM initialisation.

Add a simple AXP717 PMIC driver that knows about the buck converters, so
that we can set the voltage for the DRAM chips and the CPU cores.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>